edenapi: map TLS-related errors reported by libcurl to ApiErrorKind::Tls
Summary: When libcurl reports a TLS-related error, it is often due to a misconfiguration or certificate issue on the user's side. As such, we should treat these errors differently than other kinds of errors reported by libcurl (which are usually due to connection problems that the user has no control over). This diff introduces a new exception type for TLS errors; this will be used in a later diff to add better error messages for this kind of error.

Reviewed By: xavierd

Differential Revision: D16086607

fbshipit-source-id: 5a6425485c4326ecbc161e6ed3969041d3bc0991