edenapi: add a configurable error message for tls errors
Summary:
Similar to what was done in D15992903 for bad certificates, this diff adds a new `edenapi.tlshelp` config parameter that allows the administrator to set up a custom error message on TLS exceptions. Given that such exceptions are often caused by user misconfiguration, the error message can be set to point the user to some relevant troubleshooting steps.

This diff also changes the existing `edenapi.badcertmessage` config parameter to `edenapi.authhelp` to bring it in line with the new parameter, as well as the convention used for such configurable help messages by Commit Cloud. (This won't break anything since we have not shipped any configs that set `edenapi.badcertmessage` yet.)

Reviewed By: xavierd

Differential Revision: D16086608

fbshipit-source-id: ead83eefbd5e9bc1bf06a3aeccdc1f0737720871