remotefilelog: add iterentries to Rust datapack
Summary:
This is needed to switch `hg debugdatapack` to the Rust datapack. One major
difference with the Python code is that the Rust code builds a Vec instead of
an iterator. Implementing iterators with cpython-rust is hard (is it actually
possible?), but since the stored content isn't returned the Vec shouldn't use
too much memory.

For now, the implemention is tied to a packfile, but it shouldn't be too hard
to implement for the IndexedLogDataStore too.

Reviewed By: kulshrax

Differential Revision: D16039097

fbshipit-source-id: 567809785263bbf8cd3e1a9c24ecbb989b5c2496