mononoke: implement support for force pushrebases
Summary:
This intends to mimic the behavior of core hg server.
https://fburl.com/jf3iyl7y shows that in case when client sends a special
magic string in the `onto` field, we don't run a traditional pushrebase.
Rather, we set the rebase target to be the parent of the rebase set
(https://fburl.com/cr4ut813). The following pushrebase is just reduced to
moving a bookmark to where the `pushkey` part points.

This is equivalent to the following:
- upload all changesets (which Mononoke already does)
- run push instead of a pushrebase (e.g. move the bookmark where needed)
- generate a pushrebase-like response

NB: the code below is ugly IMO, but not more ugly than the rest of the resolver. Originally I intended to refactor the resolver, but got a pushback as my idea of refactoring was shuffling things out into different files and that would break blame.

Reviewed By: krallin

Differential Revision: D16091234

fbshipit-source-id: 6ef7f7de214edb9ac7bda14ad287a3e7f46d0014