blackbox: change session id to u64
Summary:
When considering what to do with scuba logging, we'd like some ID to chain
tables together, and is also somehow friendly to scuba.

This diff adds such an ID. It's a combination of pid (3 bytes) and the
timestamp (5 bytes). The timestamp would give scuba a change to delta-compress
the IDs. The 3-byte pid would make it possible to still see the pids using the
`blackbox` command.

Reviewed By: markbt

Differential Revision: D15755285

fbshipit-source-id: ad2dcc49514bfcd9ef474fbe0d389b9865fb2001