mononoke: remove recursion from get_hg_from_bonsai_changeset_with_impl
Summary:
Because of the recursion this function can cause stackoverflow if there are a
lot of commits that don't have hg changeset generated.
Let's rewrite it using loop_fn.

Note that there's "panic!()" in fetch_hg_changeset_from_mapping. While having
panics in the code is not great, there doesn't seem to be a way around it. When
generating new hg changesets it expects that parents' changesets are already in
the repository. If they are not generated then it's an internal logic error,
and in that case it's probably better to panic.

Reviewed By: krallin

Differential Revision: D16132403

fbshipit-source-id: 17e48ecce653f72e3f576f3698330d90a4dd7902