mononoke: lock repo from sync job itself, not from failure handler
Summary:
Sync job is used to replay pushes from Mononoke onto mercurial. It's code is in https://fburl.com/dt1hkf4g. For more about the sync job see https://fburl.com/wiki/sd2i6rch.

The goal of the sync job is to not only replicate pushes from Mononoke onto Mercurial, but also to verify that the push was correct i.e. push on Mononoke produced the same hash as on mercurial.

The problem is that we lock and unlock in two completely differen codepaths - locking is done by an external script called "failure handler" (https://fburl.com/7txm3jxf), while unlocking is a part of the sync job.

Ideally we'd like to lock and unlock in the sync job itself and remove "failure handler" entirely.

Reviewed By: StanislavGlebik

Differential Revision: D16107759

fbshipit-source-id: a418f8d0f48fa6db82476be72a91adbc03b66168