mononoke [easy]: make CensoredBlob expect ScubaBuilder rather than just table
Summary:
This means that we can later destructure and rebuild the `CensoredBlob` struct. If it is only
possible to construct it from a talbe name, there's no way to get this name
back from a destructured struct.

Note: I do this to eventually rebase D16163225 on top of this one, then be able implement `in_memory_writes_READ_DOC_COMMENT` in terms of destructuring of the existing `CensoredBlob`.

Reviewed By: krallin

Differential Revision: D16180745

fbshipit-source-id: e587675d444bf53f0bd25f01602503abad9d7d97