mononoke: allow disabling hooks from the command line
Summary:
This adds support for disabling arbitrary hooks by name using a `--disable-hook` flag. This serves as a replacement for our current approach, which is to muck with configuration when we need to disable a hook on an ad-hoc basis, with mixed results ().

There are a couple use cases for which this is useful:

- Running the Hook Tailer.
- Running Mononoke Server locally.

in both cases, you probably don't want to have the verify integrity and check_unittests hooks enabled, respectively because the former might not be installed on your server (and it'll do wrong things for hook tailing anyway), and because the latter requires a token file you're not likely to have.

Reviewed By: ahornby

Differential Revision: D16201500

fbshipit-source-id: 76404eab72b9335d82235c63003c2425f0ecc480