fix fsmonitor on Windows
Summary:
The python runtime update in D15894694 broke ctypes for all APIs that accept a
Windows `HANDLE`.  This was used by the `fsmonitor` code that tries to connect
to watchman on windows.

As a result Mercurial would silently fail to use watchman.  Commands like
`hg status` became very slow, with no user-visible output to indicate that
anything was wrong.  (It wouldn't print a progress bar or any other
information either.)

Reviewed By: quark-zju

Differential Revision: D16198869

fbshipit-source-id: 675007674e0ac742c5b62739d766b8c3016e0c1b