treemanifest: add test to catch bad tree writing
Summary:
We've seen a couple cases where hgsubversion and p4importer create
incorrect memctx's that claim to modify files without actually modifying them.
This results in some weird behavior with treemanifest when the manifest claims
to be created but results in the same hash as the parent. This breaks
historypack serialization and results in missing data in the pack.

This diff adds a test demonstrating the error. Future diffs will fix it.

Reviewed By: quark-zju

Differential Revision: D16189603

fbshipit-source-id: de6e18396dc25c70beec84e57ebdd27672a0294e