mononoke: bonsai generation: avoid content fetches
Summary:
This updates `find_file_changes` in bonsai generation to avoid content fetches. Until now, we were fetching content to see how big it is and compute its content ID, and sometimes we were also fetching the envelope to access copy info.

This updates the logic to instead unconditionally fetch the envelope, and not fetch the content at all.

Reviewed By: farnz

Differential Revision: D16262271

fbshipit-source-id: 5b299fe2769d3b8b584da2a7551434b7872396b2