procinfo: extract part of telemetry into a standalone library
Summary:
This is the first step to make the main hg executable log parent process
information to blackbox, so `hg blackbox` can show context about which parent
process triggers the hg command. It also enables us to remove the hg wrapper.

It seems the process info is the only thing that has to be done from the main
executable, instead of a separate fb-only scuba logging process. Therefore only
procinfo.rs was copied, to make it easier to build the main binary in an OSS
environment. Other information (network, OS, etc.) can remain unchanged and the
fb-only logging process can still depend on it.

This stack has breaking changes to the API exposed by procinfo. Namely, the new
APIs look the same across platforms, and does not shell out to `ps`.

Dependencies are cleaned up aggressively.

Reviewed By: sfilipco

Differential Revision: D16137796

fbshipit-source-id: 52b354a17527cd8a10d9c63d04524358c6dc3c73