mononoke: move load shedding to CoreContext
Summary:
This diff is loosely based of D15851686.

There were a few problems with the previous load shedding api. The biggest
problem was that it used static variables that required explicit
initialization. So if someone forgets to add an init in any tool (say, in
mononoke_admin), and then calls the code that runs `should_throttle` then the
binary would just crash.
The second big problem was that on each `bump_load` call configerator config
was parsing json config over and over again. I didn't measure how expensive it
was, but it looked very counter-intuitive that we put load counters in static
variables (IIRC to avoid allocations), but then we reparse the config on each
call.

To avoid these problems this diff moves ratelimiting logic into CoreContext.
The config is parsed only once on CoreContext creation and all global variables
are now local. Because of these changes `loadlimiter` library is no longer
necessary.

Note that with load limiter counters are allocated on each
`bumpLoad/shouldThrottle` call. I'll investigate how expensive it actually is.

One drawback is that now it's trickier to track egressbytes for stderr, and it was removed.
We can add it later if needed

Reviewed By: farnz

Differential Revision: D16221607

fbshipit-source-id: f934cb2fc778df1679acc7d6eb2a1a8b4deb0aac