mononoke: LFS: emit copy metadata
Summary:
We're currently not emitting any copy metadata in LFS in response to getfiles. As it happens, this appears to fine (clients still get the right metadata), because getreepack includes file copy data and so the client isn't bothered by the fact that the LFS pointers we return don't include it.

That said, it's probably a good idea to clean this up and include copy data. After all, we have it, and we should return to the client whatever they sent us.

Reviewed By: farnz

Differential Revision: D16282793

fbshipit-source-id: 6671f87de2acae2c5da1c754510bbb61de5768d3