rust: Add an inspect_result future combinator
Summary:
Add a new combinator that can be used to inspect the result of a
future. This is useful when you want to run the same code for inspect_err and inspect.

Also, update the inspect_err combinator to not panic if polled twice - instead
make it act like a fused future.

Reviewed By: farnz

Differential Revision: D16359181

fbshipit-source-id: 14948b851867d5792c76cc679297b23b1e8a6adc