types: add String/str conversions for RepoPath & friends
Summary:
In cross language code it is useful to convert to a String before giving the
object up. Without this method we would have to copy the string before dropping
the current object. This method provides an efficient method for getting the
String from RepoPath objects.
str conversions show up in similar places places.

Reviewed By: quark-zju

Differential Revision: D16352530

fbshipit-source-id: 159f522d0c19287aed29b8cce99c1675b5d801fc