avoid importing curses on Windows
Summary:
The `curses` module is part of the standard Python library on Linux and Mac,
but not on Windows.

This updates the CLI's `ui` module to avoid trying to import the `curses`
module on all platforms.  This adds a new `WindowsOutput` class for Windows.
For now this class is just a stub that uses the existing `PlainOutput` class
instead.

Reviewed By: pkaush

Differential Revision: D16354626

fbshipit-source-id: 262637030febd6893a94e19712a07cd3d5d39bbb