simplify arguments to `make-client.py`
Summary:
Update `make-client.py` so that it does not require the python interpreter to
be explicitly specified on the command line.  It now uses `sys.executable`
as the default python interpreter if one was not specified on the command
line.

Also add a `--output` flag to control the output file path.

Reviewed By: pkaush

Differential Revision: D16388625

fbshipit-source-id: 6587acb0ea8c083faf76f74c77c7c09ef0b62865