enable debug subcommands on Windows
Summary:
Enable the `edenfsctl debug` subcommands on Windows.

The `overlay` subcommand is the only debug subcommand that requires modules
not available on Windows.  The overlay code uses the `fcntl` module for
locking the overlay.  However, EdenFS on Windows does not use the overlay
directory structure used on Linux and Mac, so this subcommand isn't relevant
on Windows.

This moves the `overlay` subcommand to its own module, and changes the code to
avoid importing it on Windows.  This lets us enable the rest of the debug
subcommands on Windows.

Reviewed By: pkaush

Differential Revision: D16354620

fbshipit-source-id: ee7ea01fa3a21018850f5ba590eb143d7a9718cf