mononoke: wrap Sqlblob into a CountedBlobstore
Summary:
This seems to be the standard way (e.g. `ManifoldBlob` does it :P) to report
fb303 counters for blobstores. Provided that we want XDB Blobstore to be a
full-featured one, we want this too.

Note: potentially it makes sense to just wrap every blobstore in the `CountedBlob` in `blobstore/factory`, I can work on that later. For now I want to be able to proceed with xdb blobstore and get its counters in the cheapest possible way.

Reviewed By: krallin

Differential Revision: D16441222

fbshipit-source-id: 04fffb4598a7929301f05faacd074d5c0c76848d