censored magic string response
Summary:
Any get request for a blacklisted file returns a magic string - not an error anymore - which replaces the actual content of the file. The magic string represents a sequence of 256 random characters.
Put still returns an error if there is an attempt to change a blacklisted file, but that's the normal behavior for this type of request.

Reviewed By: ikostia

Differential Revision: D16333820

fbshipit-source-id: 54e0189ebe2a71ea365979950d44dbb3fd1e5dc7