dispatch: add friendly prefix matching on ambiguous commands
Summary:
Add support for color prefix highlighting when Mercurial encounters an ambiguous command name.  Currently ambiguous commands will print a list of possibilities such as:

```$ hg i
hg: command 'i' is ambiguous:
    id or identify
    import
    in or incoming
    init
    isbackedup
```

Now the prefix, in this case `i` will be highlighted in all possibilities.  This works no matter how long `prefix` is so in the case of `$ hg debug` it will highlight `debug` in all the possible choices.

Reviewed By: quark-zju

Differential Revision: D16451342

fbshipit-source-id: 87f58dd6bd77c2a609e5f423f7c769d444abc66d