bindings: add a wrapper to use Rust Read / Write traits on Python objects
Summary:
Add a method to wrap a Python `io` object (with `read` or `write` methods, ex.
`sys.stdin`, `ui.fout`, or `util.stringio()`) in a native Rust object so it can
be used in Rust code expecting Read / Write traits.

This will be used to expose the Rust "dispatch" logic to Python.

Reviewed By: sfilipco

Differential Revision: D16511017

fbshipit-source-id: cb1f5c20e392311cd648abe0528b971ab0266845