Commits we are going to pushrebase should be treated as drafts
Summary:
When a commit is pushrebased, we generate a new Bonsai commit to represent the result of the commit, which is the commit we actually make public.

Therefore, the commit that will be rebased should be uploaded as a draft. The rebase result will then provide filenodes, and thus we will generate good linknodes for pushrebaseed commits.

Reviewed By: krallin

Differential Revision: D16550469

fbshipit-source-id: 1f2f00f8cd9ad0f75441aca0eb8daae62ae299e0