Refactor addDelta into helper functions
Summary: Refactoring the addDelta function to allow other Journal functions to add deltas while holding the deltaState lock. This refactoring will be used to create the flush function for the journal (which needs to hold the lock to empty the journal but wants to add a delta before releasing the lock).

Reviewed By: chadaustin

Differential Revision: D16341196

fbshipit-source-id: 7b7b1d933802b466efe624378206c72c71469129