mononoke/filestore: verify hashes before committing a store
Summary:
This adds support for verifying the hashes that were provided by writers (if any) when committing a Store. This lets writers do conditional writes (i.e. if the writer knows the Sha 256 of their content, then they can ask the blobstore to verify said Sha 256 when uploading).

Note that any uploaded chunks will not be cleaned up if a conditional write fails.

Reviewed By: StanislavGlebik

Differential Revision: D16440669

fbshipit-source-id: 88bc99e646616997a4e9d7e59d59315c18f47da9