mononoke/filestore: initial implementation of the FileStore API
Summary:
NOTE: This was Jeremy's original API design for the Filestore. I've left this diff mostly unchanged (since I didn't feel super comfortable commandeering a diff then substantially changing it!), but note that the implementations for reads and writes were largely updated in my next diff, so it's probably not necessary to review them here except for context.

This is a basic implementation of the API. While the API is expressed in terms
of streaming, there's no chunking yet, either in memory or in storage.

Questions:
- Should `store` accept the aliases, or should it only take `ContentId` and always recompute the aliases?
- Should it check the ContentId is correct? It could do so easily now since its walking the entire stream anyway, but it seems like an expensive API guarantee to make.

TODO:
- Interop testing with existing blobs
- Integrate into the rest of the code
- Spawn hashers into their own tasks for parallelism
- Chunking in later diff(s)

Reviewed By: StanislavGlebik

Differential Revision: D15795778

fbshipit-source-id: e56aad086ae9e3bba0227cf3c6206faac8c97f5e